[DPEDE-2519] Fix /introduction page 404 error
